requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
langchain>=0.2.11
langchain-community>=0.2.11
langchain-core>=0.2.11
langchain-text-splitters>=0.2.0
langchain-ollama>=0.1.0
langchain-together>=0.2.0
chromadb>=0.5.3
pypdf>=4.3.1
beautifulsoup4>=4.12.3
requests>=2.32.3
sentence-transformers>=2.7.0
python-dotenv>=1.0.1
